Add Desert asset and improve UI in pages
This commit introduces a new asset named 'Desert' and exports it for usage. Additionally, improvements are made to the User Interface of several pages. These changes provide better organization and visualization of information, creating a more user-friendly application. The PlaceholderCat asset is also used for cases when no category is selected, enhancing user experience.