Add npm build run configuration and update Docker image tag
A new npm build run configuration has been added in the '.preview with compose.run.xml' file to improve the build process. Additionally, the Docker image tag has been altered in the 'Dockerfile.run.xml' file to 'dev-latest', preparing for the deployment of the new developer version of the application.