Disable parent page selector with single choice
Updated the parent page selector in the addContainer.svelte file to be disabled when there's only one choice. The 'select' HTML element now has a 'disabled' attribute that is true when there's only one or no options in the 'values' array. This change enhances user experience by preventing unnecessary and meaningless selection when only one parent page option is available.