Add new Link component and fetch favicon functionality
A new Link component was implemented, which is now used within the LinkGrid component, replacing duplicated code. We also introduced a new server-side functionality allowing the fetching of favicon and determining its dominant color for every unique URL entered. The color and favicon data received is utilized in the Link component to enhance the interface.