Optimize favicon fetch and handling
The code refines favicon fetching by first checking if the response is an image before converting it to a buffer. It also now distinguishes between different image types (icon and png) before processing data further. Additionally, a console error message has been updated to include the URL that caused an error. In the link.svelte, changes include checking that the favicon data is not none before applying styles, and adding a setup to decide if a border will be used depending on color brightness.