Correct URL validation in link form
The logic for URL checking in the link form component has been fixed. Previously the check was only passing if the URL started with 'http://', but it should also accept 'https://'. Also, 'cache' was added to the prettierignore file to prevent files in the cache directory from being automatically formatted.